# docs/public-data-api/openapi.yaml
openapi: 3.1.1
info:
  title: OWCS Public Data API
  version: 1.0.0
  description: 对外赛事数据 API 契约。旧 /agent/v1 已移除，不提供兼容转发。只提供赛事事实、名单、统计和必要的覆盖信息，不提供 AI 问答、SQL 执行、排行榜推断或系统管理数据。详见 README.md、data-dictionary.md
    与 integration.md。
servers:
- url: /data/v1
  description: 同源服务路径。
- url: https://stats.owmini.xyz/data/v1
  description: 生产服务。
security: []
...
paths:
...
  /competitions/{competition_id}/stages:
    get:
      tags:
      - 赛事
      operationId: listCompetitionStages
      summary: 获取赛事内部阶段
      description: 完整列表，按 sequence 升序。合法但尚无阶段配置的赛事返回空数组。
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageList'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    head:
      tags:
      - 赛事
      operationId: listCompetitionStagesHeaders
      summary: 获取赛事内部阶段（仅响应头）
      description: 完整列表，按 sequence 升序。合法但尚无阶段配置的赛事返回空数组。
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400Headers'
        '404':
          $ref: '#/components/responses/Error404Headers'
        '405':
          $ref: '#/components/responses/Error405Headers'
        '410':
          $ref: '#/components/responses/Error410Headers'
        '429':
          $ref: '#/components/responses/Error429Headers'
        '500':
          $ref: '#/components/responses/Error500Headers'
        '503':
          $ref: '#/components/responses/Error503Headers'
  /competitions/{competition_id}/teams:
    get:
      tags:
      - 名单
      operationId: listCompetitionTeams
      summary: 查询赛事参赛队伍
      description: 查询赛事参赛队伍
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - $ref: '#/components/parameters/q'
      - $ref: '#/components/parameters/limit'
      - $ref: '#/components/parameters/cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamList'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    head:
      tags:
      - 名单
      operationId: listCompetitionTeamsHeaders
      summary: 查询赛事参赛队伍（仅响应头）
      description: 查询赛事参赛队伍
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - $ref: '#/components/parameters/q'
      - $ref: '#/components/parameters/limit'
      - $ref: '#/components/parameters/cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400Headers'
        '404':
          $ref: '#/components/responses/Error404Headers'
        '405':
          $ref: '#/components/responses/Error405Headers'
        '410':
          $ref: '#/components/responses/Error410Headers'
        '429':
          $ref: '#/components/responses/Error429Headers'
        '500':
          $ref: '#/components/responses/Error500Headers'
        '503':
          $ref: '#/components/responses/Error503Headers'
  /competitions/{competition_id}/teams/{team_id}/players:
    get:
      tags:
      - 名单
      operationId: getCompetitionTeamRoster
      summary: 获取赛事队伍名单
      description: 完整名单，按 player.id 升序。合法参赛队伍无名单时 players=[]；未参赛队伍返回 404。名单不证明实际登场。
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - $ref: '#/components/parameters/team_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterResponse'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    head:
      tags:
      - 名单
      operationId: getCompetitionTeamRosterHeaders
      summary: 获取赛事队伍名单（仅响应头）
      description: 完整名单，按 player.id 升序。合法参赛队伍无名单时 players=[]；未参赛队伍返回 404。名单不证明实际登场。
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - $ref: '#/components/parameters/team_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400Headers'
        '404':
          $ref: '#/components/responses/Error404Headers'
        '405':
          $ref: '#/components/responses/Error405Headers'
        '410':
          $ref: '#/components/responses/Error410Headers'
        '429':
          $ref: '#/components/responses/Error429Headers'
        '500':
          $ref: '#/components/responses/Error500Headers'
        '503':
          $ref: '#/components/responses/Error503Headers'
  /competitions/{competition_id}/coverage:
    get:
      tags:
      - 赛事
      operationId: getCompetitionCoverage
      summary: 获取已收录赛事数据覆盖
      description: 获取已收录赛事数据覆盖
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - name: stage_id
        in: query
        description: 必须属于路径中的赛事；否则返回 404。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageResponse'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    head:
      tags:
      - 赛事
      operationId: getCompetitionCoverageHeaders
      summary: 获取已收录赛事数据覆盖（仅响应头）
      description: 获取已收录赛事数据覆盖
      parameters:
      - $ref: '#/components/parameters/competition_id'
      - name: stage_id
        in: query
        description: 必须属于路径中的赛事；否则返回 404。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400Headers'
        '404':
          $ref: '#/components/responses/Error404Headers'
        '405':
          $ref: '#/components/responses/Error405Headers'
        '410':
          $ref: '#/components/responses/Error410Headers'
        '429':
          $ref: '#/components/responses/Error429Headers'
        '500':
          $ref: '#/components/responses/Error500Headers'
        '503':
          $ref: '#/components/responses/Error503Headers'
...
  /matches:
    get:
      tags:
      - 比赛
      operationId: listMatches
      summary: 查询已收录系列赛
      description: 按 date 降序、id 降序。所有筛选条件为 AND；player_id 与 team_id 同时存在时，须有该选手代表该队的实际统计行；player_id 与 map_id 同时存在时，须在同一地图局统计行上满足；三者同时存在时在同一条选手地图统计上满足。用
        EXISTS 避免关联展开造成重复比赛。未知资源筛选返回 404，合法资源无结果返回空列表。
      parameters:
      - name: competition_id
        in: query
        description: 按指定资源精确筛选。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: stage_id
        in: query
        description: 仅在同时提供 competition_id 时有效；必须属于该赛事。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: team_id
        in: query
        description: 匹配比赛任一队伍。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: opponent_id
        in: query
        description: 必须与 team_id 一同使用；两者不同，匹配两队直接交手，忽略队伍侧位。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: player_id
        in: query
        description: 仅匹配至少一张地图实际有该选手统计的比赛，不按赛事名单推断出场。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: map_id
        in: query
        description: 至少有一张该地图的比赛；使用 EXISTS 筛选，不重复返回系列赛。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: date_from
        in: query
        description: 比赛日期下界，包含当天。
        schema:
          type: string
          format: date
      - name: date_to
        in: query
        description: 比赛日期上界，包含当天；不得早于 date_from。
        schema:
          type: string
          format: date
      - $ref: '#/components/parameters/limit'
      - $ref: '#/components/parameters/cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchList'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    head:
      tags:
      - 比赛
      operationId: listMatchesHeaders
      summary: 查询已收录系列赛（仅响应头）
      description: 按 date 降序、id 降序。所有筛选条件为 AND；player_id 与 team_id 同时存在时，须有该选手代表该队的实际统计行；player_id 与 map_id 同时存在时，须在同一地图局统计行上满足；三者同时存在时在同一条选手地图统计上满足。用
        EXISTS 避免关联展开造成重复比赛。未知资源筛选返回 404，合法资源无结果返回空列表。
      parameters:
      - name: competition_id
        in: query
        description: 按指定资源精确筛选。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: stage_id
        in: query
        description: 仅在同时提供 competition_id 时有效；必须属于该赛事。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: team_id
        in: query
        description: 匹配比赛任一队伍。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: opponent_id
        in: query
        description: 必须与 team_id 一同使用；两者不同，匹配两队直接交手，忽略队伍侧位。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: player_id
        in: query
        description: 仅匹配至少一张地图实际有该选手统计的比赛，不按赛事名单推断出场。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: map_id
        in: query
        description: 至少有一张该地图的比赛；使用 EXISTS 筛选，不重复返回系列赛。
        schema:
          type: integer
          minimum: 1
          description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
      - name: date_from
        in: query
        description: 比赛日期下界，包含当天。
        schema:
          type: string
          format: date
      - name: date_to
        in: query
        description: 比赛日期上界，包含当天；不得早于 date_from。
        schema:
          type: string
          format: date
      - $ref: '#/components/parameters/limit'
      - $ref: '#/components/parameters/cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400Headers'
        '404':
          $ref: '#/components/responses/Error404Headers'
        '405':
          $ref: '#/components/responses/Error405Headers'
        '410':
          $ref: '#/components/responses/Error410Headers'
        '429':
          $ref: '#/components/responses/Error429Headers'
        '500':
          $ref: '#/components/responses/Error500Headers'
        '503':
          $ref: '#/components/responses/Error503Headers'
  /matches/{match_id}:
    get:
      tags:
      - 比赛
      operationId: getMatch
      summary: 获取系列赛
      description: 获取系列赛
      parameters:
      - $ref: '#/components/parameters/match_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
              example:
                data:
                  id: 5445
                  competition:
                    id: 24
                    name: 2026 守望先锋世界杯 小组赛
                  stage: null
                  date: '2026-08-23'
                  format: BO5
                  team1:
                    team:
                      id: 53
                      name: AUS
                    score: 3
                  team2:
                    team:
                      id: 65
                      name: MEX
                    score: 2
                  winner_team_id: 53
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    head:
      tags:
      - 比赛
      operationId: getMatchHeaders
      summary: 获取系列赛（仅响应头）
      description: 获取系列赛
      parameters:
      - $ref: '#/components/parameters/match_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400Headers'
        '404':
          $ref: '#/components/responses/Error404Headers'
        '405':
          $ref: '#/components/responses/Error405Headers'
        '410':
          $ref: '#/components/responses/Error410Headers'
        '429':
          $ref: '#/components/responses/Error429Headers'
        '500':
          $ref: '#/components/responses/Error500Headers'
        '503':
          $ref: '#/components/responses/Error503Headers'
  /matches/{match_id}/games:
    get:
      tags:
      - 地图与统计
      operationId: listMatchGames
      summary: 获取比赛地图局
      description: 完整返回。number 已知项按 number、id 升序，未知项最后按 id 升序。ID 排序不代表真实比赛顺序。
      parameters:
      - $ref: '#/components/parameters/match_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameList'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    head:
      tags:
      - 地图与统计
      operationId: listMatchGamesHeaders
      summary: 获取比赛地图局（仅响应头）
      description: 完整返回。number 已知项按 number、id 升序，未知项最后按 id 升序。ID 排序不代表真实比赛顺序。
      parameters:
      - $ref: '#/components/parameters/match_id'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: 成功。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '304':
          description: 当前公开表示未改变；无响应正文。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '400':
          $ref: '#/components/responses/Error400Headers'
        '404':
          $ref: '#/components/responses/Error404Headers'
        '405':
          $ref: '#/components/responses/Error405Headers'
        '410':
          $ref: '#/components/responses/Error410Headers'
        '429':
          $ref: '#/components/responses/Error429Headers'
        '500':
          $ref: '#/components/responses/Error500Headers'
        '503':
          $ref: '#/components/responses/Error503Headers'
...
components:
  schemas:
...
    RosterResponse:
      type: object
      additionalProperties: false
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/Roster'
    CoverageResponse:
      type: object
      additionalProperties: false
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/Coverage'
...
    TeamList:
      type: object
      additionalProperties: false
      required:
      - data
      - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Team'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    StageList:
      type: object
      additionalProperties: false
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Stage'
          description: 完整返回该父资源范围，不分页。
...
  parameters:
    competition_id:
      name: competition_id
      in: path
      required: true
      description: 资源 ID；父子资源不匹配返回 404。
      schema:
        type: integer
        minimum: 1
        description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
    team_id:
      name: team_id
      in: path
      required: true
      description: 资源 ID；父子资源不匹配返回 404。
      schema:
        type: integer
        minimum: 1
        description: 当前数据集内稳定的赛事资源标识；不得按名称代替 ID，也不得假设跨环境 ID 一致。
...
    limit:
      name: limit
      in: query
      description: 每页条数。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    cursor:
      name: cursor
      in: query
      description: 沿用上一页全部筛选参数和 limit，原样传入下一页令牌。令牌绑定查询条件与有效期，不保证跨页数据库快照。
      schema:
        type: string
        maxLength: 4096
    q:
      name: q
      in: query
      description: 名称包含查询串，忽略大小写并规范空白；% 和 _ 按普通字符，不是通配符；规范后空字符串无效。按 ID 返回全部候选。队伍同时检索已确认 aliases；不自动合并同名实体。
      schema:
        type: string
        minLength: 1
        maxLength: 100
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: 上一次相同请求的 ETag，用于条件 GET/HEAD。
      schema:
        type: string
  headers:
    ETag:
      description: 公开响应表示的验证器，涵盖所有公开子数据；不能只用父比赛的更新时间。
      schema:
        type: string
    Cache-Control:
      description: 普通赛事查询使用 public, max-age=0, must-revalidate。缓存复用前须验证。
      schema:
        type: string
    X-Request-Id:
      description: 仅用于问题定位的请求标识，放在响应头，不进入赛事对象。
      schema:
        type: string
  responses:
    Error400:
      description: 无效参数、未知参数、无效游标或游标与筛选条件不匹配。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error400Headers:
      description: 无效参数、未知参数、无效游标或游标与筛选条件不匹配。
    Error404:
      description: 资源不存在、父子不匹配或队伍不属于指定赛事。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404Headers:
      description: 资源不存在、父子不匹配或队伍不属于指定赛事。
    Error405:
      description: 只允许 GET 和 HEAD。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Allow:
          description: 允许的赛事读取方法。
          schema:
            type: string
            enum:
            - GET, HEAD
    Error405Headers:
      description: 只允许 GET 和 HEAD。
      headers:
        Allow:
          description: 允许的赛事读取方法。
          schema:
            type: string
            enum:
            - GET, HEAD
    Error410:
      description: 列表游标已过期；从首页重新读取。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error410Headers:
      description: 列表游标已过期；从首页重新读取。
    Error429:
      description: 请求过多。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          description: 重试等待秒数。
          schema:
            type: integer
            minimum: 1
    Error429Headers:
      description: 请求过多。
      headers:
        Retry-After:
          description: 重试等待秒数。
          schema:
            type: integer
            minimum: 1
    Error500:
      description: 服务器内部错误。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error500Headers:
      description: 服务器内部错误。
    Error503:
      description: 来源不可用或该响应无法满足数据完整性约束。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error503Headers:
      description: 来源不可用或该响应无法满足数据完整性约束。